+++ inlang/packages/paraglide/paraglide-js/CHANGELOG.md
@@ -4,6 +4,15 @@
 
 ## 2.0.0-beta.27
 
+- fix wrong matching in API requests [#427](https://github.com/opral/inlang-paraglide-js/issues/427)
+
+Paraglide JS is no longer extracting the locale from API requests for the `url` strategy because that can lead to unwanted re-directs. To get the right locale in API requests, at least add the `baseLocale` strategy to your options. 
+
+```diff
+-strategy: ["url"]
++strategy: ["url", "cookie", "baseLocale"]
+```
+
 - `experimentalMiddlewareLocaleSplitting` option https://github.com/opral/inlang-paraglide-js/issues/425#issuecomment-2692351073
 
 - fix [setLocale() triggers re-loads if the same locale is set](https://github.com/opral/inlang-paraglide-js/issues/430)